pyproject.toml
[tool]

# ==== Poetry ====
[tool.poetry]
name = "frogbase"
version = "2.0.0a4"
description = "FrogBase simplifies the download-transcribe-embed-index workflow for multi-media content. It does so by linking content from various platforms with speech-to-text models, image & text encoders and embedding stores."
homepage = "https://hayabhay.github.io/frogbase"
repository = "https://github.com/hayabhay/frogbase"
documentation = "https://hayabhay.github.io/frogbase"
authors = ["Abhay Kashyap"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
packages = [
    { include = "frogbase" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.urls]
Changelog = "https://github.com/hayabhay/frogbase/releases"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<=3.11.4"
yt-dlp = "^2023.3.4"
loguru = "^0.7.0"
tinydb = "^4.8.0"
humanize = "^4.6.0"
webvtt-py = "^0.4.6"
pydantic = "^2.0"
torch = ">=2.0.0, !=2.0.1"
openai-whisper = "^20230314"
sentence-transformers = "^2.2.2"
hnswlib = "^0.7.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
tox = "^4.6.2"
pre-commit = "^3.3.3"
virtualenv = "^20.23.1"
pip = "^23.1.2"
twine = "^4.0.2"
toml = "^0.10.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.3.2"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
flake8-docstrings = "^1.7.0"
pytest-cov = "^4.1.0"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocs-include-markdown-plugin = "^4.0.4"
mkdocs-material = "^9.1.16"
mkdocs-video = "^1.5.0"
mkdocstrings = "^0.22.0"
mkdocs-autorefs = "^0.4.1"
mike = "^1.1.2"


[tool.poetry.group.ui]
optional = true

[tool.poetry.group.ui.dependencies]
streamlit = "^1.23.1"
python-dotenv = "^1.0.0"
openai = "^0.27.8"

[tool.poetry.scripts]
frogbase = "frogbase.__main__:ui"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# ==== Coverage ====
[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["frogbase", "tests"]


# ==== black ====
[tool.black]
line-length = 119
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# ==== isort ====
[tool.isort]
profile = "black"
line_length = 119


# ==== mypy ====
[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
